refactor: MemberAcceptanceTest 패키지 이동
