+++ src/test/java/nextstep/subway/favorite/acceptance/FavoriteAcceptanceTest.java
@@ -11,7 +11,7 @@ import nextstep.subway.favorite.dto.FavoriteResponse;
 import nextstep.subway.line.acceptance.LineAcceptanceTest;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
-import nextstep.subway.member.MemberAcceptanceTest;
+import nextstep.subway.member.acceptance.MemberAcceptanceTest;
 import nextstep.subway.station.StationAcceptanceTest;
 import nextstep.subway.station.dto.StationResponse;
 import org.junit.jupiter.api.BeforeEach;
